requirements.txt
matplotlib
pandas
requests
flask
python-dotenv
dash
Flask-Caching
